Anadrome String
If the all letter of a given string occour even times, or only one letter occours odd times, then the input string is anadrome. For example: aann is anadrome (only "even occouring"), "zanna" is anadrome too, because only one letter (in this case 'z') occours odd times, but "zeanna" is not a anadrome string (as well 'z' as 'e'occours odd times). The program tests if the given string is anadrome.